Move download_data to ismi.io
